+++ src/main/java/net/malisis/core/mixin/core/MixinChunkCollision.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package net.malisis.core.asm.mixin.core;
+package net.malisis.core.mixin.core;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.spongepowered.asm.mixin.Mixin;
