+++ src/main/java/net/malisis/core/mixin/core/MixinClientNotif.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package net.malisis.core.asm.mixin.core;
+package net.malisis.core.mixin.core;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
