+++ src/main/java/net/malisis/core/mixin/core/client/MixinMalisisRenderer.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package net.malisis.core.asm.mixin.core.client;
+package net.malisis.core.mixin.core.client;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
