+++ src/main/resources/mixins.malisiscore.core.json
@@ -1,7 +1,7 @@
 {
     "required": true,
     "minVersion": "0.7.4",
-    "package": "net.malisis.core.asm.mixin.core",
+    "package": "net.malisis.core.mixin.core",
     "refmap": "mixins.malisiscore.refmap.json",
     "target": "@env(DEFAULT)",
     "compatibilityLevel": "JAVA_8",
